Move certificate expiry check method to common module
