+++ open-banking-accelerator/components/com.wso2.openbanking.accelerator.common/src/main/java/com/wso2/openbanking/accelerator/common/util/CertificateUtils.java
@@ -82,4 +82,23 @@ public class CertificateUtils {
         // remove spaces, \r, \\r, \n, \\n, ], [ characters from certificate string
         return value.replaceAll("\\\\r|\\\\n|\\r|\\n|\\[|]| ", StringUtils.EMPTY);
     }
+
+    /**
+     * Check whether the certificate is expired.
+     *
+     * @param peerCertificate the certificate to be checked
+     * @return true if the certificate is expired
+     */
+    public static boolean isExpired(X509Certificate peerCertificate) {
+        try {
+            peerCertificate.checkValidity();
+        } catch (CertificateException e) {
+            log.error("Certificate with the serial number " +
+                    peerCertificate.getSerialNumber() + " issued by the CA " +
+                    peerCertificate.getIssuerDN().toString() + " is expired. Caused by, " + e.getMessage());
+            return true;
+        }
+        return false;
+    }
+
 }
