+++ open-banking-accelerator/components/com.wso2.openbanking.accelerator.gateway/src/main/java/com/wso2/openbanking/accelerator/gateway/executor/util/CertificateValidationUtils.java
@@ -71,18 +71,6 @@ public class CertificateValidationUtils {
         // Adding a private constructor to hide the implicit public one.
     }
 
-    public static boolean isExpired(X509Certificate peerCertificate) {
-        try {
-            peerCertificate.checkValidity();
-        } catch (CertificateException e) {
-            LOG.error("Certificate with the serial number " +
-                    peerCertificate.getSerialNumber() + " issued by the CA " +
-                    peerCertificate.getIssuerDN().toString() + " is expired. Caused by, " + e.getMessage());
-            return true;
-        }
-        return false;
-    }
-
     /**
      * Get issuer certificate from the truststore.
      *
